fixes #294 Change alerts tools to use email address from profile
Tool has also been moved to /usr/local/topaz/scripts directory as a script.
It does run a bit slower - but as just one query (instead of 2 that are
manually merged).

I tried to borrow the design from 
  source:head/packages/tools/src/main/resources/scripts/annotated.groovy
which to me seems the easiest to read/maintain structure to date.


git-svn-id: http://svn.topazproject.org/svn/head@2393 48088309-b40f-0410-b149-b14953253675