Provide mappings to rdf:List, rdf:Bag, rdf:Seq and rdf:Alt.
Note that there is no consensus on annotations and defaults used etc. 
So those things might change. But for now if the triples representation
of a field need to be any of the container types or collection type as
defined in the Rdf specs, an explicit annotation is required.

Also note that the MapperType concept introduced is a bit experimental and
may change later.



git-svn-id: http://svn.topazproject.org/svn/head@2534 48088309-b40f-0410-b149-b14953253675