Generate test-classes.jar so post-integration doesn't remove all traces
This is desirable if you ever want to access the compiled test classes after
the tests are run. In my case, I'm trying to use the compiled otm samples.

(Integration test results are removed in an attempt to prevent maven from
using them for unit-tests... not that something in maven still causes things
to be run more than once sometimes.)


git-svn-id: http://svn.topazproject.org/svn/head@2553 48088309-b40f-0410-b149-b14953253675