Collections API as summarized on wiki:ArticleCollections
Ronald and Pradeep were correct that Criteria provides most of what is
needed. The SimpleCollection and SmartCollection classes provided here
are really just for convenience to formalize the desired pattern. They
do simplify a few things though and may be worth using when collections
are really implemented.

The more interesting bits that need to be implemented are additional
mulgara resolvers.


git-svn-id: http://svn.topazproject.org/svn/head@2603 48088309-b40f-0410-b149-b14953253675