Added development/debugging helper to avoid the need to run a CAS server. To use, uncomment the filter-mapping for DummySSOFilter in web.xml. Then you can log in/out and create new users simply by appending "sso.auth.id" and "sso.email" query parameters to any request.
Example usage:

Log in with auth-id 'foo':
{{{
    http://localhost:8080/plosone-webapp/home.action?sso.auth.id=foo&sso.email=a@b.com
}}}
If this is the first time, a new user will be created (you'll be asked to set
up the profile etc. Now going to
{{{
    http://localhost:8080/plosone-webapp/home.action
}}}
will leave you still logged in (i.e. you navigate the site as normal). To log
out, go to (empty auth-id)
{{{
    http://localhost:8080/plosone-webapp/home.action?sso.auth.id=&sso.email=
}}}
or to immediately log out and in again with auth-id 'bar' go to
{{{
    http://localhost:8080/plosone-webapp/home.action?sso.auth.id=bar&sso.email=bar@b.com
}}}


git-svn-id: http://svn.topazproject.org/svn/head@2717 48088309-b40f-0410-b149-b14953253675