OTM-ize the annotation and reply services.
Currently the AnnotationInfo and ReplyInfo wrappers are
used to transfer the values from otm-model classes
so that there is least impact. An issue with directly
using the otm-models is that this would also requre
modifying the java-scripts. This is because the
action.* classes are serialized by the JSON marshaler
and sent across to the annotation/reply java-scripts.

Note also that the /admin/adminTop.action is untested
because of a mulgara dead-lock caused by actualizing
a lazy loaded UserProfile object outside of a txn scope.



git-svn-id: http://svn.topazproject.org/svn/head@2727 48088309-b40f-0410-b149-b14953253675