customize WebWorks to allow the mapping of arbitrary URIs to arbitary WebWorks actions and allow a clean result path where application is responsible for HTTP headers & content. - PlosOneActionMapper is now the default WebWorks URI mapper   - it currently filters for URIs that start with "/article/feed" and sends them to the getFeed action (ArticleFeed) - PlosOneFeedResult is a WebWorks result that understands Atom feeds
with this pattern, it is possible to create human (or machine) friendly request URIs that do not follow WebWorks templating rules or processing.
it is also possible to create WebWorks results that do not go through WebWorks templating or processing.
as these are still WebWorks actions, they execute in the existing web frameworks for security, caching, logging, etc.


git-svn-id: http://svn.topazproject.org/svn/head@2819 48088309-b40f-0410-b149-b14953253675