re #349 Mulgara-service to use plos WebAppListener (broken by [2830])
Unless we move mulgara-service to plos, this is an example of where
we'd like to have a more generic config singleton that doesn't belong
to either topaz or plos.

Not sure what happens to the mulgara-service though when we move to
RMI. We probably wouldn't use the WebAppListener concept then and
would instead just have to write a few lines of code to manually
create the commons-config stuff. So the org.plosone:conf-helper
dependency (that seems weird) can be removed at that point for this
module.


git-svn-id: http://svn.topazproject.org/svn/head@2832 48088309-b40f-0410-b149-b14953253675