Make sure that any proxy object returned by the session.get() is loaded from store and not just a stub.
This was causing a problem when criteria/oql calls
with cross associations return a resultset some of
the objects may not be loaded from the store - 
contrary to the expectation of the user.

The problem surfaced after converting criteria to
make use of session cache since the PlosOne reply service
had cross associations.



git-svn-id: http://svn.topazproject.org/svn/head@2893 48088309-b40f-0410-b149-b14953253675