Fix a bug in Session.get() where a lazy loaded instance is ignored and a new instance is created under certain cases. This can happen when the instance to be loaded has associations that contain references to this instance (back-pointer) and the setter for this association is triggering a load of the lazy loaded association stub (eg. if the collection is a HashSet)
The fix is to make sure to always use the instance that the users
have seen (via a setter method) by always attempting to merge the
newly loaded object with the ones in cache - even though such a check
is done before the load.

This can in some cases cause a load of the same object from the store
twice. Therefore the copy() merge is optimized to recognize proxy
objects and so the merge is simply to switch the backing object.

To support this optimization, the concept of pristine-proxies has
been changed. Session now keeps track of all proxy objects it 
generates. Also a sub-class of MethodHandler is used to obtain the 
backing store state information.

This fixes #369 where ingest() was failing because Article object 
had a duplicate pristine proxy instance and hence prevented
the write of the articleState.



git-svn-id: http://svn.topazproject.org/svn/head@2920 48088309-b40f-0410-b149-b14953253675