Remove xacml from search service
Next changes will be to add xacml back into plosone where it calls search.
This will be a bit more work since results will have to be cached on
plosone, etc. Until these changes are made, search will return unpublished
articles in its result set.

I think there may still be some xacml/pep stuff in the search infrastrcture
(as it was simply a template copied from the other services), but none of
it is used and it can be removed -- just not sure how to do that properly.


git-svn-id: http://svn.topazproject.org/svn/head@2950 48088309-b40f-0410-b149-b14953253675