Use the Spring ApplicationContext singleton to look up Spring instantiated beans in code that is not under Spring's domain.
Reverted r2992 completely. Reworked r2997 and r3000 to get the OTM
session from Spring. Spring magic works here too. The OTM session 
returned is in the correct scope as defined in Spring config file.
(in this case the applicationContext.xml)

This also means r2989 for OTM is not needed for PlosOne webapp
(and likely for any Spring applications). May be useful for
other applications and hence leaving that alone. Note that calling
getCurrentSession() from within the PlosOne webapp will fail
since the otm SessionFactory used by PlosOne webapp is no longer
configured with a CurrentSessionContext manager.

Thanks Ronald for pointing to the Spring singletons.



git-svn-id: http://svn.topazproject.org/svn/head@3029 48088309-b40f-0410-b149-b14953253675