re #340 Move spring stuff into its own namespace
The commons-config stuff used as placeholders in spring now exist
inside <plosone><spring> in commons-config. I'm not completely sure
I like adding the <spring> layer, but there are going to be several
checkins moving around config in an attempt not to break anything
between my testing and other's changsets. And deeper namespaces are
better than no-namespaces.


git-svn-id: http://svn.topazproject.org/svn/head@3045 48088309-b40f-0410-b149-b14953253675