Share the ItqlHelper handle from otm transaction in ArticleUtil.delete().
ArticleUtil is shared with command line tools and it is unaware of 
otm-models and SessionFactory. Instead of adding all that support for 
command line tools now into ArticleUtil, we should have an integrated 
plosone.jar (as has been discussed before) that can be a more useful lib 
than this.

Also the getObjectURL() method in ArticleUtil is no longer called from 
plosone app. Instead a static method is added to create the fedora
REST URL for accessing a representation.

With this, all usages of ItqlHelper in plosone is tied to an OTM 
transaction and hence we are ready to do the 'Open View in Session'
(See http://www.hibernate.org/43.html).



git-svn-id: http://svn.topazproject.org/svn/head@3090 48088309-b40f-0410-b149-b14953253675