re #340 global_config.ftl now uses common-config
plos-configuration's WebAppListener now adds an attribute to the
servlet context containing common-config's root Configuration
object. Freemarker is able to see this attribute as an application
scope variable and use it to retrieve configuration.

The specific configuration parameters needed are the location of
the plosone webapp. These needed to be added to the cas.xml and
the defaults.xml that backs it.


git-svn-id: http://svn.topazproject.org/svn/head@3092 48088309-b40f-0410-b149-b14953253675