re #70 Initial pass at script to extract more data from fedora
This is just a first pass that actually works so that feedback on the
results this script produces can be gathered. Of particular interest
is what data I'm moving, the actual predicates and whehter they all
belong in Article or some belong in ObjectInfo, etc.

Example Usage:
  /usr/local/topaz/bin/rungroovy /usr/local/topaz/scripts/migration.groovy -a info:doi/10.1371/journal.pone.0000056

You can then go into itql and do the following:
{{{
/usr/local/topaz/bin/runitql
%mode = "table reduce exp quote"
%trunc = 50
select $p $o from <local:///topazproject#ri> where <info:doi/10.1371/journal.pone.0000056> $p $o;
}}}

You should see things like:
{{{
<topaz:pageCount>          5
<topaz:copyrightYear>      2006
<topaz:issue>              1
<topaz:volume>             1
<topaz:articleType>        'research-article'
<topaz:copyrightStatement> 'Vigne, Frelin. This is an open-access article di...'
<topaz:publisherName>      'Public Library of Science'
<topaz:journalTitle>       'PLoS ONE'
<topaz:authors>            _node483
<topaz:affiliations>       'Institut National de la SantÃ© et de la Recherche...'
<topaz:affiliations>       'UniversitÃ© de Nice Sophia Antipolis'
<topaz:body>               '<body><sec id=\'s1\'><title>Introduction</title><p...'
52 rows
}}}

The intent is to use the fedora-client to iteratate over all the articles it knows
about. That is something like the following lists all the articles in fedora:
{{{
$FEDORA_HOME/client/bin/fedora-find localhost 9090 pid '?' http | grep pid | grep 'pone.[0-9]*$' | awk '{print $2}'
}}}
If the output from that is looped around this migration.groovy script, all articles
will get migrated.

IMO, The migration script itself should leave around some kind of predicate
indicating what version of the migration script was run. i.e. it would be
a version number that we'd not want the actual plosone application to use.


git-svn-id: http://svn.topazproject.org/svn/head@3108 48088309-b40f-0410-b149-b14953253675