fixes #346 Add article references to mulgara
Created new Reference/Citation models and added them to Article.
migration script adds data appropriately to mulgara.

The references frequently leave off data or have malformed data.
That is, values that are integers are not always there. Thus, the
data type is Integer (not int). And, worse, sometimes these integers
are actually strings -- in which case the migration tool logs a
warning and behaves as if the value did not even exist -- thus no
triple is created. I don't think it is a frequent occurance, but
it would be good to check /var/log/topaz/tools.log after a major
run to be sure.


git-svn-id: http://svn.topazproject.org/svn/head@3136 48088309-b40f-0410-b149-b14953253675