Created an embedded DublinCore class capture the existing DC elements of ObjectInfo and Article. Unfortunately this requires a new triple to be added to all 'Objects' (the ObjectInfo and Article class) in existing stores. This triple is rdf:type rdf:resource="http://rdf.plos.org/RDF/Object.
I have tested ingesting a new article and can browse it. Unfortunately the
VirtualMapper goes into infinite recursion when I try and retrieve it. I have
sent an email to Jeff on this. I am going to go ahead the check this in, but
suspect a few problems might still crop up.

Feedback appreciated.



git-svn-id: http://svn.topazproject.org/svn/head@3147 48088309-b40f-0410-b149-b14953253675