re #463 Don't show number of search hits until we've reached the end
Doing so previously wasn't a big deal because the only reason we'd
be off was xacml filtering related to an unpublished article. (And
unpublished articles were relatively rare in the system.) For virtual
journals, however, lucene's equivalent of count(*) could frequently
be very different from the number of results we return after filtering.
Thus, we only return the string "of N results" on the last page.
(Further optimization could be done if the user goes backward to
remember the number of filtered results, but there seem to be other
priorities at the moment.)


git-svn-id: http://svn.topazproject.org/svn/head@3167 48088309-b40f-0410-b149-b14953253675