Defined the Aggregation model. Thanks everyone for the feed-back on the RFC.
Note that the 'hasPart' used for the static collection is commented out
since that conflicts with the 'hasPart' in DublinCore. Will be moving that
out from DublinCore and uncommenting this very soon (as per a discussion
with Amit)

Also note that the smart-collection rules are defined in terms of
OTM DetachedCriteria for now. This could change if that is too limiting.




git-svn-id: http://svn.topazproject.org/svn/head@3187 48088309-b40f-0410-b149-b14953253675