re #486 Allow manual setting of self-link and make id unique via URN
Plos will be manually pulling these feeds and making them available via
apache. Thus, the need to manually set the self-link. To do so, simply
add the selfLink cgi-parameter. For example:
 * /article/feed?selfLink=http://www.plosone.org/rss/articles.xml
 * /article/feed?category=Biotechnology&selfLink=http://www.plosone.org/rss/categories/biotechnology.xml

Also, feedburner insists on a valid globally-unique feed id. So using
the URN I believe we've registered, simply set the ID to
  info:doi/10.1371/feed.pone

This can be set via pub.feed.id in plosone.xml. And it should probably
be changed for different journals as the feeds will be different.

An initial attempt to keep this id unique for different categories and
authors is done by adding cgi-parameters to the URN, but this should be
revisited as feed support is expanded. I could find very little on the
net indicating when doing this kind of thing is appropriate. My
suggestion is to view the id as a cache key and make sure it changes
if anything other than new elements end up in the feed.


git-svn-id: http://svn.topazproject.org/svn/head@3211 48088309-b40f-0410-b149-b14953253675