flesh out Clinical Trials config and overrides - created struts-clinicalTrials.xml for CT Action overrides - added placeholder action + content for CT home page - added placeholder CT favicon, MRI of Homer Simpson's brain
note that mappingPrefix was moved back to VirtualJournalContextFilter from VirtualJournalMappingFilter.
the "mappingPrefix" can be thought of as the "namespace" to look up resources.
the ActionMapper also needs to know this namespace so it should be set as part of the context.
it is expected that there may be more instances of use to come.


git-svn-id: http://svn.topazproject.org/svn/head@3215 48088309-b40f-0410-b149-b14953253675