Add trackback functionality re #74.  Couple of notes:
Since there has been no consensus on URLs, the current ones are the 
default struts2 URLs.  The trackback url that must be entered on 
a blog site is of the form:

http://localhost:8080/plosone-webapp/trackback.action?trackbackId=info%3Adoi%2F10.1371%2Fjournal.pone.0000010

The code will check for the existence of a previous entry from the same
blog URL and will ignore subsequent requests.  It will also get the blog
entry and make sure there is a link to the article from the blog.  It will
use the journalContext object to figure out the correct URL to look for.
Finally, the code does not currently keep a different count for different
journals, i.e. if the article appears in both CT and PLoSONE, no distinction
is made in the trackback tally. 



git-svn-id: http://svn.topazproject.org/svn/head@3277 48088309-b40f-0410-b149-b14953253675