fixes #362 deleteannotation script
Note that presently this deletes annotations, discussions and ratings.
However, there is a bug in the OTM and I'm not clear if all 3 of these
types of annotations will continue to be deleted with this script.

Also, this does not clear the plosone cache. Thus, to see the result
of this script, plosone must be restarted.


git-svn-id: http://svn.topazproject.org/svn/head@3281 48088309-b40f-0410-b149-b14953253675