Make sure all access to parent or child member fields are via a getter/setter method call in DetachedCriteria.
This was causing problems with lazy-loaded proxies of
DetachedCriteria because the proxy can't detect 
member variable access and hence a load was not performed.

Also make use of rdf:Seq for the various ordered lists
now that r3248 fixed the type-look-ahead issue for
rdf:Seq.

Also added a test for DetachedCriteria persistance.



git-svn-id: http://svn.topazproject.org/svn/head@3288 48088309-b40f-0410-b149-b14953253675