fixes #332 Add volume and issue number to extended ATOM feed
Note that you must specify the extended feed to get the issue and volume.
That is, the cgi parameters should include extended=true.

The namespace for this extension is as indicated in #332 and is configurable
in plosone.xml (see diffs associated with this checkin).

For example, with article 440, the following will be added to its entry:
{{{
    <plos:volume xmlns:plos="http://www.plos.org/atom/ns#plos">2</plos:volume>
    <plos:issue xmlns:plos="http://www.plos.org/atom/ns#plos">5</plos:issue>
}}}


This same mechanism can probably be used for all ATOM extensions. And to
clarify what that is (given the rome api intentionally under-documents
this in an attempt to prevent people from arbitrarily extending ATOM), the
interface is as follows:
  com.sun.syndication.feed.atom.Entry.setForeignMarkup(List<org.jdom.Element>);

See [http://www.jdom.org/docs/apidocs/index.html jdom api] - specifically
javadocs for Element. I'm fairly certain these things can be nested, but
haven't tried yet.


git-svn-id: http://svn.topazproject.org/svn/head@3296 48088309-b40f-0410-b149-b14953253675