Integration test frame work for plosone.
run it using 'mvn -Pit-helper clean install'. 

First time will be slow. Known issues:

  * launcher process does not die for plosone and mulgara

Multiple environments can be configured and installed (See Env.groovy)
Currently there is a one-to-ome correspondence between the data-set
and the environment. Data-set is a tar-gz of the following tree: 
{{{
   data
   `-- fedora
   |   |-- objects
   |   |-- datastreams
   `-- mulgara  (mulgara data dir)
}}}

Environment install is only done once. To upgrade, delete the state
file named 'installed' from the install directory. Since the install
step is expensive, the install directory should not be kept under the 
'target' directory so that 'mvn clean' does not blow it away.


git-svn-id: http://svn.topazproject.org/svn/head@3333 48088309-b40f-0410-b149-b14953253675