fixes #329 Adds subject atom feed to browse by subject pages
NOTE: I'm not making canonical category names by the mechanism indicated in
ticket #329 because there is no way to go the other way. And when we start
using dynamic feeds in production (after #469 perhaps), we need the real
category name to be passed to the feed action. Instead, I'm URL encoding
the category name so spaces and single quotes will become + and %27
respectively. If static files are served, apache will have to deal with that.
I talked to Russ about this on the phone.

There are four new configuration elements under feed. The defaults are for
/article/feed, but plosone.xml is setup to still use the static feeds.

plosone.xml configures static feeds as:
 * /rss/PLoSONE.xml
 * /rss/Genetics+and+Genomics ''(note it does NOT end with .xml)''

If the new settings in plosone.xml are removed, the feeds will default to:
 * /article/feed
 * /article/feed?category=Genetics+and+Genomics

The default for feeds is to show the last 3-months of articles. However, this
might not be deisreable for categories (subjects). If the category (subject)
feeds want to show all articles since the beginning of time, change
<pub><feed><categoryPrefix> to "feed?startDate=&category=". Doing so would
change the feed URL for categories to:
 * /article/feed?startDate=&category=Genetics+and+Genomics

(This capability was added in r3363.)


git-svn-id: http://svn.topazproject.org/svn/head@3365 48088309-b40f-0410-b149-b14953253675