fixes #570 Fix timing on feed caches
The cache keys were being defined including the start-date. And the start-
date was defined as 3-months prior to the current timestamp including
hours, minutes and seconds. So, everytime a new feed was requested, the
start-date was changing (unless the request was in the same second). To
fix this, we now zero out the hours, minutes and seconds.

This fix also makes the number of months back configurable. (Jeff and I
debated whether this should really be in days or weeks and decided to
leave it in months for now.)

Add a bit of logging.


git-svn-id: http://svn.topazproject.org/svn/head@3536 48088309-b40f-0410-b149-b14953253675