Couple of updates:   * Added Support for CT, PLOSONE journals.   * Tests for viewing articles in CT and PLOSONE
Also renamed. HomePageTest to NavTest. A test class per page is good when 
running with -Dtest=xxx - but that would mean a full run will have to 
start/stop environment for every page - and that is pretty slow. Also
this allows simulation of gotoPage() keeping the current browser context.
Faster than beginAt() since that starts a new browser simulation context.
Also the gotoPage() will be a better simulation of user actions.

Also a minor update: envs[] is no longer protected. Use getBasic() 
and getEmpty() to get the server environment for the test.



git-svn-id: http://svn.topazproject.org/svn/head@3616 48088309-b40f-0410-b149-b14953253675