Changes made to release.properties for release:prepare
There are still a vew missing artifacts under project.dev that cause the
end of release:prepare to fail (after the tag has been created). Of course,
it probably isn't critical in the branch itself unless there is active work
in the branch. But somebody will want to ensure all the artifacts are there
(and perhaps prune the old ones) when a release is done from head next.


git-svn-id: http://svn.topazproject.org/svn/branches/0.8@3700 48088309-b40f-0410-b149-b14953253675