Merge in r3763 from head:
Reduce the HttpSession timeout for mulgara to 10 minutes.

This is because there is 1 to 1 correspondence between an HttpSession and
the itql-interpreter-bean and under certain error conditions
(eg. SocketTimeOutException) the axis soap-rpc client seems to be loosing
the http-session cookie. This would mean that now there is no way for the
client code to close the the itql-interpreter-bean. It'll be closed only
on an HttpSession? timeout. So to minimize the effect of the lost session,
the session time out is reduced to 10 minutes to match the SocketTimeout?
value.

The conditions described above are observed from the log files related
to #650.

The real solution is to either explicitly pass a handle
in the client/server API or use RMI as suggested by #645.



git-svn-id: http://svn.topazproject.org/svn/branches/0.8@3771 48088309-b40f-0410-b149-b14953253675