Merged in [3780] from head to address #662. Original commit message was:
Move otm-transaction scope from around whole results.getPage() to around each
individual article check in the HitGuard. While this isn't as efficient, it
prevents mulgara from holding the write lock for, in some cases, several
minutes - this is most notably the case when lucene takes several seconds for
a search and the request is directly for something like the 10th page (i.e.
100 search hits have to be loaded). Best would be to wrap the transaction
around the checking of a page full of articles, but that doesn't work well
with the Iterator based structure used here.

This addresses #662.


git-svn-id: http://svn.topazproject.org/svn/branches/0.8@3781 48088309-b40f-0410-b149-b14953253675