Replaced manual transaction management via TransactionHelper with declarative transaction management using otm's spring transaction manager. This has the advantage of simplifying the code and allowing transaction boundaries to be set and controlled in a central place (the spring config). The disadvantage is some hacks to stop/start transactions in the code (more below).
Spring has a number of methods to control transaction boundaries - see
http://forum.springframework.org/showthread.php?t=37178 for some discussion
of these. Since we're using spring 2.0, the transaction-advice and the
@Transactional annotation are the main candidates for us. While @Transactional
has the advantage of putting the transaction declaration near the code that
needs them (though they must be placed on a public method that's invoked from
the outside, as class-internal calls do not get intercepted), the disadvantage
is closer tying to spring and a more fragmented transaction management.
Furthermore, one of goals of this change was to once-and-for-all get rid of
the lazy-load exceptions and avoid the touch-all-objects hacks. Hence I opted
for using aop tx-advice injection in the spring config.

This approach works well for most things, but there are few exceptions:

 * creating user-accounts requires two transactions in order to ensure
   uniqueness of the account-name. This requires a hack where the code commits
   and starts a new otm transaction explicitly using otherwise hidden methods
   in the otm transaction-manager.
 * because lucene queries can take a while and because mulgara currently
   supports only a single transaction at a time, we need to apply the same
   hack as above in search in order to disable the transaction during the
   lucene query. (see also issue #662)
 * the journal service registers a callback with ehcache, and this callback
   may be invoked either within the context of a web-request or within the
   context of an RMI call (on changes in caches on peers). I did not see a
   clean way of handling this via spring and the transaction-manager without
   a fair amount of infrastructure, and hence opted to stay with
   TransactionHelper here.
 * permissions are initialized from a webapp-context-listener, which is outside
   the scope of spring, and hence needs to use TransactionHelper to manage the
   transaction needed during this initialization.

Another potential problem is that transactions are now longer (the duration of
the action), which means less concurrency. Due to the fact that otm starts the
underlying mulgara transaction lazily (i.e. only when an actual db call is
needed), this isn't quite as bad as it may initial sound - e.g. requests that
just hit the caches fully concurrent. However, ingest now has an even larger
impact (the image-scaling really needs to be moved out now).

Lastly, separate transactions are used for the action executions and for each
getter on the action (i.e. the results accesses by the freemarker template).
This is partly due to the fact that we cannot intercept the request at a high
enough level in spring (AFAIK), and partly to avoid keeping the transaction
open too long (see preceding paragraph). The downside is A) extra transactions
(but only if lazy-loading is being triggered, which is seldom the case now and
will hopefully be even less the case in the future) and B) a slight potential
for inconsistent data in the UI (again, only if lazy loading is being
triggered by the UI).


git-svn-id: http://svn.topazproject.org/svn/head@4040 48088309-b40f-0410-b149-b14953253675