Fixed article-retrieval, broken by [4040].
When chaining actions all properties on the previous action are copied to the
next action so that they're available to the view. The ChainingInterceptor
does this by invoking OgnlUtil.copy() which copies everything it can from the
source to the target object, where "it can" means there's a getter on the
source and a setter on the target (or direct field access). What was happening
was that this was also copying various properties introduced by the
transaction interceptor, most especially the "targetSource" and "callbacks"
properties - the effect was that the tx-interceptor for the target class
(TrackbackAction) was now suddenly trying to invoke its methods on the source
class (FetchArticleAction), leading to a ClassCastException.

This fix makes sure we exclude certain properties from being copied. This is
somewhat fragile in that a new version of spring could introduce new
properties that need to be also excluded; however, the alternative is a list
of includes which is much longer and which also needs to be maintained. The
fix is further complicated by the fact that for some reason the
ChainingInterceptor's include and exclude lists are Collection's but there's
no way to specify a collection in the struts.xml config - hence the type
converter and conversion-properties file.


git-svn-id: http://svn.topazproject.org/svn/head@4113 48088309-b40f-0410-b149-b14953253675