Made object-state cache shared. The state cache is now managed using weak-references to the object itself and when a detached object is re-attached to another session, the new session now knows the last persisted state of that object.
Note that this does not change the behavior of Session.get().
It will only return objects from the session cache or will 
create a new instance. So the management of changes made via 
duplicate instances is still an application's responsibility. 
But because of change-track and partial updates based on states
maintained globally, the application may see a different
behavior than prior to this check-in. See the changes to 
testCascade() in this check-in for such an example.



git-svn-id: http://svn.topazproject.org/svn/head@4116 48088309-b40f-0410-b149-b14953253675