Tightened handling of objects in 'Removed' state.
An object was deleted but was still referenced by another object is the case
that is handled here. The previous behavior attempts to do an un-delete. 
This was mainly done since we had implicit orphan-deletes before. Now that 
orphan-delete is done only when an application explicitly requests it, there 
is no need to do the auto-undelete and in fact it overrides the 
'strong intent' by the app via a delete() call or the explicit 'orphan-delete' 
directive.

A bonus is that the Object state documentation from Hibernate/JPA is now
applicable verbatim for OTM.



git-svn-id: http://svn.topazproject.org/svn/head@4141 48088309-b40f-0410-b149-b14953253675