Replaced aging ItqlHelper and Answer objects with new, shinier ones. The new mulgara-client now has an ItqlClient interface together with several implementations for soap, rmi, and embedded db access; an ItqlClientFactory is used to create instances (this is currently quite primitive and needs to be enhanced at some point to allow the setting of parameters such as authentication info for soap client and database location for embedded db). Additionally the Answer object has been simplified: Answer is an interface, with appropriate implementations for the different clients.
With this it is now possible to talk to an unmodified mulgara via rmi by
specifying an "rmi:..." server uri; it also possible to run with an embedded
mulgara by specifying a "local:..." server uri.

In order to avoid pulling in 25MB of jars when not needed, the dependencies for
the different client impls are marked optional in the pom. In order to use the
soap client, axis must be explicitly be added as a dependency to the apps pom;
similarly, to use the rmi client the mulgara driver jar must be added, and to
use the embedded db the full mulgara jar is needed.

Lastly, the driver jar dependency in the mulgara-client pom has currently been
completely commented out because of problems building due to that jar including
an older version of axis.


git-svn-id: http://svn.topazproject.org/svn/head@4179 48088309-b40f-0410-b149-b14953253675