Implement lazy loading for collections of associations. All that is lazily done is the creation of the collection containing associations obtained via session.load(). No changes to the ITQL query and therefore the association-ids and the type-look-ahead information is still loaded.
However this lazy load has the advantage that java.util.Set impls
don't force load the association. This fixes #627. Now at least a 
collection's semantics need not be compromised for performance. ie.
we can now use java.util.Set where that is needed instead of the
java.util.List that was used for performance reasons.

Also see the notes in 
http://lists.topazproject.org/pipermail/topaz-dev/2008-January/000063.html 



git-svn-id: http://svn.topazproject.org/svn/head@4605 48088309-b40f-0410-b149-b14953253675