fixes #479 - Extended the DOI Resolver to map annotation DOIs to the plosone server. The supported doi URL should be formatted as:
http://doi-resolver-host:port/web-app-context-path/10.1371/annotation/xxxxx

where "annotation/xxxxx" above is the part of the annotation DOI <info:doi/annotation/xxxxx>. Note that the word "annotation" need not be present, but it is part of the actual DOI we typically assign to annotations. 

In this case, the DOIResolver makes no assumptions about the given annotation DOI format (unlike the prior implementation for article DOI's). The DOI of an annotation contains no reference to the journal, so the DOIResolver must query the root annotated item (an article?), and will use the journal regular expressions already defined to figure out which plos journal to redirect the annotation request to. Ultimately we rely on the article doi containing a reference to the journal, which is not ideal. We need a more reliable way to determine the correct journal to redirect to in a future implementation. This could be fixed when the whole journal-filtering implementation is re-visited. 

git-svn-id: http://svn.topazproject.org/svn/branches/0.8.2.1@4660 48088309-b40f-0410-b149-b14953253675