I was going to check all this in one go, when I realized that subversion does not deal too well with modifications and rename/moves at the same time. I will check this in as an intermediate (complilable and works), move to 'extras' and then clean-up.
I have moved away from Eric's idea of creating OTM objects to generate OWL
triples for the simple reason as to how much work do we want to duplicate. We
still might have to do it in the future, but will wait till the content model
gets fleshed out a bit. For the time being we will make use of 
[http://owlapi.sourceforge.net/ OWLAPI] to generate the OWL ontology from Topaz
class meta-data. Please note: OWLAPI, while making use of maven, does not
publish to a repository so I have copied the jars to our internal maven
repository. I have informed the OWLAPI team and they are looking into it.

The design essentially is driven by a single file OwlGenerator which extracts
the requisite information from OTM and adds them as 'axioms' to OWLAPI.


git-svn-id: http://svn.topazproject.org/svn/head@4673 48088309-b40f-0410-b149-b14953253675