fixes #770 Less than ideal fix for intermittent HTTP 302 server error msg when attempting to invoke startDiscussion.action.
Added check for existence of 'userAttributeKey' in http session before presenting the URL linking to start discussion UI form.
This "check" is already existed for certain other links in the RHC and I merely replicated this for the "Make a general comment" link.  
The downside to this fix is that a secure redirect to CAS will happen when a user is not "properly" logged in to the current journal that subsequently re-directs back to the article.
(NOTE: This was and is the behavior for the "Add a note to the text" link in the RHC)

Notes/history for future ideal fix:
I was able to repeat this error consistently (prior to the fix) on my machine by performing the following in order:
1. Clear all plos related cookies
2. Goto a NON-PlosONE virtual journal and login THERE
3. Goto a PlosONE journal article.
5. Attempt to submit a comment via clicking the Add a general comment link in the RHC
6. 302 error occurrs.

A better fix would seem to be making the virtual journal boundaries more transparent with respect to a logged in users credentials and associated tokens so the check for 'userAttributeKey' w/in the FTL context would be un-necessary.
I'm sure the server side Filters could do this.


git-svn-id: http://svn.topazproject.org/svn/branches/0.8.2.1@4724 48088309-b40f-0410-b149-b14953253675