Closes #763: add JTA support to OTM. The current approach takes a TM and makes it available to parties who are then responsible for enlisting themselves (aka their resource-managers). The TM is set on the SessionFactory; if none is set then (for local transactions) a JOTM instance is used. Transactions can either be local (started by Session.beginTransaction()) or global (managed via normal JTA mechanisms).
Internally things have been moved away from passing a transaction object around
to using a connection instead: the first time a store is accessed during a
transaction a connection to that store is opened and this connection is used
for all operations until the transaction is complete at which point it is
closed. Stores will typically want to enlist their XAResource at the time the
connection is opened; delisting is done automatically.

The ItqlStore properly uses Mulgara's XAResource; the FedoraBlobStore is faking
things currently.

One thing not properly supported yet is transaction recovery, as this is TM
implementation dependent.

The local transaction interface has been enhanced in two ways: there's an option
to start a read-only transaction (Session.beginTransaction(true)), and a
transaction may now be marked for rollback-only (Transaction.setRollbackOnly()).
Read-only transactions are extremely useful with the current Mulgara which
supports an arbitrary number of read-only transactions but only a single
write-transaction at any given time.


git-svn-id: http://svn.topazproject.org/svn/head@4807 48088309-b40f-0410-b149-b14953253675